# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Hotel Las Vegas Reservation Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Room {
  id: number;
  number: string;
  type: string;
  floor: number;
  status: string;
  pricePerNight: number;
  /** @nullable */
  description?: string | null;
  createdAt: string;
}

export interface RoomInput {
  number: string;
  type: string;
  floor: number;
  pricePerNight: number;
  description?: string;
}

export interface RoomUpdate {
  number?: string;
  type?: string;
  floor?: number;
  status?: string;
  pricePerNight?: number;
  description?: string;
}

export interface Employee {
  id: number;
  name: string;
  role: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  createdAt: string;
}

export interface EmployeeInput {
  name: string;
  role: string;
  phone: string;
  email?: string;
}

export interface EmployeeUpdate {
  name?: string;
  role?: string;
  phone?: string;
  email?: string;
}

export interface Guest {
  id: number;
  name: string;
  phone: string;
  nationalId: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  nationality?: string | null;
  createdAt: string;
}

export interface GuestInput {
  name: string;
  phone: string;
  nationalId: string;
  email?: string;
  nationality?: string;
}

export interface GuestUpdate {
  name?: string;
  phone?: string;
  nationalId?: string;
  email?: string;
  nationality?: string;
}

export interface ReservationDetail {
  id: number;
  roomId: number;
  guestId: number;
  employeeId: number;
  checkInDate: string;
  checkOutDate: string;
  status: string;
  totalAmount: number;
  paymentReceiptNumber: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  room: Room;
  guest: Guest;
  employee: Employee;
}

export interface ReservationInput {
  roomId: number;
  guestId: number;
  employeeId: number;
  checkInDate: string;
  checkOutDate: string;
  totalAmount?: number;
  paymentReceiptNumber: string;
  notes?: string;
}

export interface ReservationUpdate {
  roomId?: number;
  guestId?: number;
  employeeId?: number;
  checkInDate?: string;
  checkOutDate?: string;
  totalAmount?: number;
  paymentReceiptNumber?: string;
  notes?: string;
  status?: string;
}

export interface DashboardStats {
  totalRooms: number;
  occupiedRooms: number;
  availableRooms: number;
  reservedRooms: number;
  totalReservations: number;
  activeReservations: number;
  todayCheckIns: number;
  todayCheckOuts: number;
  monthlyRevenue: number;
  totalRevenue: number;
}

export interface OccupancyItem {
  type: string;
  total: number;
  occupied: number;
  available: number;
}

export interface EmployeeStatItem {
  employeeId: number;
  employeeName: string;
  role: string;
  totalReservations: number;
  activeReservations: number;
}

export type ListRoomsParams = {
status?: string;
type?: string;
};

export type ListGuestsParams = {
search?: string;
};

export type ListReservationsParams = {
status?: string;
employeeId?: number;
roomId?: number;
checkInDate?: string;
checkOutDate?: string;
};

export type ExportReservationsCSVParams = {
status?: string;
startDate?: string;
endDate?: string;
};