# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hotel Las Vegas Reservation Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List all employees
      responses:
        "200":
          description: List of employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create an employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "404":
          description: Not found
    patch:
      operationId: updateEmployee
      tags: [employees]
      summary: Update an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeUpdate"
      responses:
        "200":
          description: Updated employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "404":
          description: Not found
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Delete an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /guests:
    get:
      operationId: listGuests
      tags: [guests]
      summary: List all guests
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of guests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Guest"
    post:
      operationId: createGuest
      tags: [guests]
      summary: Create a guest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuestInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guest"
  /guests/{id}:
    get:
      operationId: getGuest
      tags: [guests]
      summary: Get a guest
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Guest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guest"
        "404":
          description: Not found
    patch:
      operationId: updateGuest
      tags: [guests]
      summary: Update a guest
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuestUpdate"
      responses:
        "200":
          description: Updated guest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guest"
        "404":
          description: Not found
    delete:
      operationId: deleteGuest
      tags: [guests]
      summary: Delete a guest
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
components:
  schemas:
...
    Employee:
      type: object
      required: [id, name, role, phone, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        createdAt:
          type: string
    EmployeeInput:
      type: object
      required: [name, role, phone]
      properties:
        name:
          type: string
        role:
          type: string
        phone:
          type: string
        email:
          type: string
    EmployeeUpdate:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
        phone:
          type: string
        email:
          type: string
    Guest:
      type: object
      required: [id, name, phone, nationalId, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        nationalId:
          type: string
        email:
          type: ["string", "null"]
        nationality:
          type: ["string", "null"]
        createdAt:
          type: string
    GuestInput:
      type: object
      required: [name, phone, nationalId]
      properties:
        name:
          type: string
        phone:
          type: string
        nationalId:
          type: string
        email:
          type: string
        nationality:
          type: string
    GuestUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        nationalId:
          type: string
        email:
          type: string
        nationality:
          type: string
...