# lib/api-zod/src/generated/types/employee.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Hotel Las Vegas Reservation Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface Employee {
  id: number;
  name: string;
  role: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  createdAt: string;
}