# lib/api-zod/src/generated/types/employeeInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Hotel Las Vegas Reservation Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface EmployeeInput {
  name: string;
  role: string;
  phone: string;
  email?: string;
}