# lib/api-zod/src/generated/types/guestInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Hotel Las Vegas Reservation Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface GuestInput {
  name: string;
  phone: string;
  nationalId: string;
  email?: string;
  nationality?: string;
}